requirements.txt
langchain
langchain-google-genai
google-generativeai
python-dotenv
fastapi
uvicorn
streamlit
requests



# Core
python-dotenv
logging

# FastAPI (backend API)
fastapi
uvicorn

# LangChain (framework)
langchain
langchain-openai

# OpenAI SDK (used internally by langchain-openai)
openai

# Optional: If you want Ollama fallback (local free models)
langchain-community
